Listo calisto

+++ La_Rata_Tostada_se_va_por_las_Ramas/Assets/Scripts/Minigames/Demons_Controller.cs
@@ -43,7 +43,7 @@ public class Demons_Controller : IMinigame
         reset = false;
     }
 
-    public override int CalculateScore()
+    public override float CalculateScore()
     {
         return points;
     }
