+++ La_Rata_Tostada_se_va_por_las_Ramas/Assets/Scripts/Minigames/Popups.cs
@@ -71,7 +71,7 @@ public class Popups : IMinigame
         reset = false;
     }
 
-    public override int CalculateScore()
+    public override float CalculateScore()
     {
         return points;
     }
