+++ La_Rata_Tostada_se_va_por_las_Ramas/Assets/Scripts/Minigames/Powerpoint.cs
@@ -102,7 +102,7 @@ public class Powerpoint : IMinigame
         reset = false;
     }
 
-    public override int CalculateScore()
+    public override float CalculateScore()
     {
         return points;
     }
